1. Naming chef zero provisoner to chef infra, 2. switching default provisoner in kitchen from solo to zero(now infra),  3. adding provisoner test for chef infra and chef zero backward compatibility
