+++ README.md
@@ -83,7 +83,7 @@ More verbose logging for test-kitchen can be specified when running test-kitchen
 $ kitchen test -l debug
 ```
 
-For the provisioner (e.g. chef-solo or chef-zero) add a `log_level` item to the provisioner section of the `.kitchen.yml`
+For the provisioner (e.g. chef-solo or chef-infra) add a `log_level` item to the provisioner section of the `.kitchen.yml`
 For more information on setting `log_level` see the [configuration documentation](https://kitchen.ci/docs/reference/configuration/).
 
 ## Documentation
